modify: fetch를 페이지 내 컴포넌트가 아닌 page.tsx에서 하도록 수정(리렌더링 방지)
