+++ app/user/shop/components/BottomSheetContent.tsx
@@ -1,10 +1,8 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 'use client';
 
 import { GiftDto } from '@/application/usecases/gift/dto/GiftDto';
 import GiftListItem from '@/components/GiftListItem';
 import EventBus from '@/types/EventBus';
-import { useEffect, useState } from 'react';
 import styled from 'styled-components';
 
 /* ---------------------------------- style --------------------------------- */
@@ -26,46 +24,22 @@ const BSContentList = styled.ul`
 type BottomSheetContentProps = {
   setSelectedItemKey: (key: string) => void;
   moveSheetToBottom: () => void;
+  giftList: GiftDto[] | null;
+  loading: boolean;
 };
 
 /* ---------------------------------- component --------------------------------- */
 export default function BottomSheetContent({
   setSelectedItemKey,
   moveSheetToBottom,
+  giftList,
+  loading,
 }: BottomSheetContentProps) {
   const handleItemClick = (key: string) => {
     setSelectedItemKey(key);
     EventBus.emit('itemClicked', true);
     moveSheetToBottom();
   };
-  const [giftList, setGiftList] = useState<GiftDto[] | null>(null);
-  const [loading, setLoading] = useState<boolean>(true);
-
-  useEffect(() => {
-    const getGifts = async (): Promise<any> => {
-      try {
-        setLoading(true);
-        // disabled 되지 않은 gift list 가져오는 api 사용
-        const res = await fetch('/api/user/shop', { method: 'GET' });
-        const result = await res.json();
-
-        if (!result.success) {
-          console.error('API 오류:', result.message);
-          setGiftList([]);
-          return [];
-        }
-        console.log('가져옴', result.data);
-        setGiftList(result.data);
-      } catch (error) {
-        console.error('데이터 로드 실패:', error);
-        return [];
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    getGifts();
-  }, []);
 
   return (
     <BSContentContainer>
